sks ta: factorize cipher/authen sequences
This change mainly aims are simplifying integration of
asymmetric cryptography without putting a mess in the
implementation.

Symmetric operations through the GPD TEE Core Internal API
are implementation in processing_symm.c. The few very specific
AES things remain in processing_aes.c.

Entry points for almost all steps of encrypt, decrypt, sign
and verify functions are factorize: all use the same sequence
and get references to an input and an output data buffer
(reference may be NULL). Only the oneshot and the update steps
for the verify function differ in that they gets 2 input data
memory reference parameters instead of an input and an output
one.

This change also modifies management of the session processing
state:
  - Remove check_processing_state(). Simply check the session
    field @processing instead.
  - Introduce reset_processing_state() to reset it to ready
    state.
  - set_processing_state() gets the target function to process
    instead of a target processing state. The function also
    gets 2 object references since those used in the processing
    may impact the processing state as for the need of user
    authentication.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>