C_GetAttributeValue implementation and other improvements
Signed-off-by: Gabor Szekely <szvgabor@gmail.com>

- Implemented C_GetAttributeValue command, but access control needs
to be implemented before this can be used

- Modified code related to C_FindObjectsInit and friends,
so it will match all objects if there are no attributes supplied
in the template, as specified by the standard.

- Added SKS_CKFT_LOGIN_REQUIRED to token flags as default,
because otherwise openssh wouldn't ask for pin number and fail,
since the private key is only available to authenticated users.
I'm not 100% sure that this is the right way to add this flag,
comments are welcome.

- Added automatic CKA_ID generation to the generate key pair
code, because otherwise openssh wouldn't use the generated
key pair. The standard mentions that public and private key
objects may be paired based on their id (although doesn't
require this to be enforced), I assume this is why openssh
only works with key pairs that have CKA_ID.