Add TEE_ALG_RSASSA_PKCS1_V1_5 algorithm
Signed-off-by: Gabor Szekely <szvgabor@gmail.com>

Note that TEE_ALG_RSASSA_PKCS1_V1_5 is not part of the GPD TEE
standard. It was added, because I couldn't find a way to implement
the CKM_RSA_PKCS mechanism signing and verifying through the TEE
API. The TEE_ALG_RSASSA_PKCS1_V1_5 currently uses EMSA padding and
signs the message without hashing it.