scripts: sign_encrypt.py: use keys from a PKCS#11 token
Adds argument --pkcs11-sign to sign TA with a private key stored
in a PKCS#11 token. The PKCS#11 token is used to compute the RSA
signature of the image digest.

PKCS#11 target token and key are defined from environmental variables
or can be statically defined in the python script.
- token label: PKCS11_TOKEN_LABEL, defaults to 'OP-TEE TA authentication'
- token user PIN: PKCS11_TOKEN_PIN, defaults to None (token not logged in)
- key label: PKCS11_KEY_LABEL, defaults to None (don't match by CKA_LABEL)
- key ID: PKCS11_KEY_ID, defaults to None (don't match by CKA_ID)

To sign TAs with the private key:
  sign_encrypt.py pkcs11-sign --uuid XXXXXXXXXXXXXXXX

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>