tree-wide: use ROUND{UP|DOWN}2() where applicable
Use ROUNDUP2() and ROUNDDOWN2() where the rounding size argument
is not a constant value. This change prepares a later change where
ROUNDUP() and ROUNDDOWN()) macros will only support constant size
argument values but will be able to ensure that value is a power of 2
without impacting the implementation in size or efficiency.

This change uses ROUNDUP2() and ROUNDDOWN2() instead of ROUNDUP_VAR()
and ROUNDDOWN_VAR() since the former less increase the resulting
binary image in non-debug mode, as compiler nicely optimize the
mult and div operations on such power of 2 constant size argument values.

Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>