tree-wide: use ROUNDUP_OVERFLOW_VAR() where applicable
Use ROUNDUP_OVERFLOW_VAR() where the rounding size argument is not a
constant value. This change prepares a later change where
ROUNDUP_OVERFLOW() macro will only support constant size argument
values but will be able to ensure that value is a power of 2
without impacting the implementation in size or efficiency.

Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>